+++ src/server.c
@@ -461,14 +461,14 @@ void server_init (struct audio_s *audio_config_p, struct misc_config_s *mc)
  * This waits for a client to connect and sets an available client_sock[n].
  */
 #if __WIN32__
-	connect_listen_th = (HANDLE)_beginthreadex (NULL, 0, connect_listen_thread, (void *)(unsigned int)server_port, 0, NULL);
+	connect_listen_th = (HANDLE)_beginthreadex (NULL, 0, connect_listen_thread, (void *)(ptrdiff_t)server_port, 0, NULL);
 	if (connect_listen_th == NULL) {
 	  text_color_set(DW_COLOR_ERROR);
 	  dw_printf ("Could not create AGW connect listening thread\n");
 	  return;
 	}
 #else
-	e = pthread_create (&connect_listen_tid, NULL, connect_listen_thread, (void *)(long)server_port);
+	e = pthread_create (&connect_listen_tid, NULL, connect_listen_thread, (void *)(ptrdiff_t)server_port);
 	if (e != 0) {
 	  text_color_set(DW_COLOR_ERROR);
 	  perror("Could not create AGW connect listening thread");
@@ -484,14 +484,14 @@ void server_init (struct audio_s *audio_config_p, struct misc_config_s *mc)
 	for (client = 0; client < MAX_NET_CLIENTS; client++) {
 
 #if __WIN32__
-	  cmd_listen_th[client] = (HANDLE)_beginthreadex (NULL, 0, cmd_listen_thread, (void*)client, 0, NULL);
+	  cmd_listen_th[client] = (HANDLE)_beginthreadex (NULL, 0, cmd_listen_thread, (void*)(ptrdiff_t)client, 0, NULL);
 	  if (cmd_listen_th[client] == NULL) {
 	    text_color_set(DW_COLOR_ERROR);
 	    dw_printf ("Could not create AGW command listening thread for client %d\n", client);
 	    return;
 	  }
 #else
-	  e = pthread_create (&cmd_listen_tid[client], NULL, cmd_listen_thread, (void *)(long)client);
+	  e = pthread_create (&cmd_listen_tid[client], NULL, cmd_listen_thread, (void *)(ptrdiff_t)client);
 	  if (e != 0) {
 	    text_color_set(DW_COLOR_ERROR);
 	    dw_printf ("Could not create AGW command listening thread for client %d\n", client);
@@ -535,10 +535,10 @@ static THREAD_F connect_listen_thread (void *arg)
 	SOCKET listen_sock;  
 	WSADATA wsadata;
 
-	snprintf (server_port_str, sizeof(server_port_str), "%d", (int)(long)arg);
+	snprintf (server_port_str, sizeof(server_port_str), "%d", (int)(ptrdiff_t)arg);
 #if DEBUG
 	text_color_set(DW_COLOR_DEBUG);
-        dw_printf ("DEBUG: serverport = %d = '%s'\n", (int)(long)arg, server_port_str);
+        dw_printf ("DEBUG: serverport = %d = '%s'\n", (int)(ptrdiff_t)arg, server_port_str);
 #endif
 	err = WSAStartup (MAKEWORD(2,2), &wsadata);
 	if (err != 0) {
@@ -658,7 +658,7 @@ static THREAD_F connect_listen_thread (void *arg)
 
     	struct sockaddr_in sockaddr; /* Internet socket address stuct */
     	socklen_t sockaddr_size = sizeof(struct sockaddr_in);
-	int server_port = (int)(long)arg;
+	int server_port = (int)(ptrdiff_t)arg;
 	int listen_sock;  
 	int bcopt = 1;
 
@@ -1265,7 +1265,7 @@ static THREAD_F cmd_listen_thread (void *arg)
 					/* Maximum for 'D': Info part length + 1 */
 	} cmd;
 
-	int client = (int)(long)arg;
+	int client = (int)(ptrdiff_t)arg;
 
 	assert (client >= 0 && client < MAX_NET_CLIENTS);
